Preliminary support for invoking DRES from actions (similar to $(MAKE)).
execute_actions has now hardcoded support for invoking dres_update_goal
as an action. It is hardocded as the dres() action right now. It is quite
useless, as we do not have dres-level (&-) variables implemented yet. To
make this useful we need to supported nested scoping for dres-variables.

dres_update_goal