Fix graph generation to push also nodes that never had incoming edges.
This bug was triggerable with a ruleset that resulted in a unary tree
(ie. a fully 'linear' graph) where the last node had no dependencies.
Eg. for the simplest example below the graph generation algorithm failed
to include 'bar' in the graph. As a result it gave a trivially false
positive on circular dependencies.

foo: bar

bar:
    echo(i_am_severly_broken)