Implemented some missing assignment and local variable support.
Added support for immediate assignments ($foo.bar = 1) and local variable
assignments. ATM it is kludgish at best, waiting for the big cleanup /
refactoring / rewrite. Once most of the functionality is in place and we see
all the needed runtime semantics of global and local variables, we can clean
up the grammar (make it more regular/symmetric among other things) and clean
up / refactor / reimplement the affected part of the code.