Added fallback action handler support.
It is possible now to register a fallback action handler by calling
dres_fallback_handler. If such a handler is registered all unknown actions
are passed to this handler, ie. whenever an action with no registered handler
is to be executed DRES will call the fallback handler. If no such handler
is registered, unknown actions result in a runtime error.