Topologically sorted depenedency graphs are now built only once.
The dependency graph is now built and sorted for each and every target
once. Currently this happens automagically on-demand during the first
call to dres_update_goal. I think we'll need to expose an explicit
interface for this as well so that the caller can catch errors in the
input file before/without trying to update any goals.