Create/commit/rolll back transactions only at the top-level DRES invokation.
Nested factstore transactions seem to be broken (or have very strange
semantics). As a workaround we generate and handle transactions only at
the outermost invokation (even in the case of targets that recursively
invoke DRES).

Timestamp updating is still broken: we don't/cannot undo timestamp updates
in case of rollbacks causing subsequent goal resolution to be potentially
incorrect/incomplete.