Preliminary support for saving/loading binary rule files.
We can now save the parsed input file to and load from a binary format.
The binary format does not contain all information (on purpose), only
the minimum to run the resolver. This excludes the parsed per-target
actions. Only the byte code resulting from compiling these is retained.
We might (need to) add support for optionally saving the actions as well
for debugging.

There were a couple of changes during the implementation of binary saving /
loading. Most notably the name/ID mapping of the registered methods/action
handlers is now saved and restored. The generated byte code now uses the IDs
to refer to methods instead of strings (both are supported by the VM).

Generally speaking the saved binary should be endianness-agnostic as we
save everything in network byte order. The current exception is the byte
code which needs some fixing. The byte code sometimes uses 16 (and IIRC even
24) bit integers within the instruction stream (eg. to denote the length of
a string embedded in the instruction stream). Currently these are not saved
in an endianness agnostic manner (ie. end up always in host byte order). Needs
to be fixed, might require dedicated support in the VM code (vm_chunk_save /
vm_chunk_load).

There are a couple of currently unsolved problems binary saving / loading
shed a light on. These are mostly related to how the various initialization /
finalization functions are distributed between dres_init and dres_parse_file
(and to the fact that dres_load for a binary file is equivalent to dres_init +
dres_parse_file + dres_finalize for a source file). Needs some careful checking
and reshuffling of the code.