Allow integers and doubles as locals for dres_update_goal.
Added an ugly hack that allows one to pass also integer and double
arguments to dres_update_goal/push_locals without breaking backward-
compatibility. The old interface allowed string-only name/value pairs
as locals. The current hack checks whether value looks to be a char
tag ('s', 'i', or 'd') and if so it takes and interprets the next
argument accordingly (string, integer, double). If it does not look
to be a tag, it is assumed to be a call for the old interface hence
value is used as a string. Yuck, once we merge/update everyone
(mostly libplayback backend) depending on us I'll remove the backward-
compatibility support and require all arguments from all callers to
ba tagged (but with "s", "i", and "d" instead of 's', 'i', and 'd').