Added support for partial updates of facts using the operator |=.
The VM UPDATE instruction has been extended to support partial updates and
the compiler has been changed to emit partial VM UPDATE instructions for
the operator "|=".

A partial update (vm_fact_update), as opposed to a full update (vm_fact_copy)
does not touch any field of the destination operand (fact) that is not present
in the source operand (fact). Moreover, updates to fields with identical
values are suppressed. As a result partially updating a fact with a subset
(including the full set) of its current field values does not touch the
fact at all (and thus avoids the factstore signaling a change/update).

Note that the partial assignment operator "|=" is valid only for updates,
ie. assignemnts of the form "$fact[field1, ... fieldn] = ...". For non-update
assignments it is difficult to come up with sensible semantics for a partial
assignment. One could be to avoid overwriting any fact with a fully identical
copy of itself to minimize factstore update signals but that can be trivially
accomplished by using a partial update instead ($fact[key(s)] |= ...)