Changed return value conventions of action handlers to support (silent) failure.
Action handlers now return values < 0 for errors. This causes the VM to raise
an exception and DRES to abort updating the goal and rolling back the current
transaction. A return value of FALSE (== 0) causes the transaction to be
rolled back without any exceptions. Any other return values (> 0) are inter-
preted as TRUE and cause the goal update processing to continue normally.