Added support for multiply-defined/split targets.
Targets can now be defined in multiple fragments. The multiple
fragments are combined to a single definition by concatenating
both the prerequisits and statements in the order they are
encountered in the input. Multiple references to the same object
in the prerequisits are reduced to the first occurence. Eg.

all: foo bar
	echo('all #0...')

all: bar foobar
	echo('all #1...')

all: bar barfoo foo
	echo('all #2...')

is equivalent to

all: foo bar foobar barfoo
	echo('all #0...')
	echo('all #1...')
	echo('all #2...')