Implemented a new replacing assignment operator '*='.
Using this new operator it is now possible to add/remove facts and
replace a subset of facts that match a given selection criteria.
The syntax of the operator is

<$fact><optional_selector> *= <expression>, where

fact is a valid fact name, optional_selector - if given - is a selector
identical to the one used in partial assignments and update operators,
and expression is any expression producing a set of facts.

Some more concrete examples:

Replace all instances of $foo with the ones returned by predicate xyzzy:

  $foo *= prolog(xyzzy, arguments...)

Update matching instances of $foo (using field bar for matching) with the
ones returned by predicate xyzzy, append the rest of returned facts as
new ones, and remove any other existing instances of $foo:

  $foo[bar] *= prolog(xyzzy, arguments...)

Same as above but apply it only to those instances of $foo for which
field foobar != 3:

  $foo[foobar:!3,bar] *= prolog(xyzzy, arguments...)

For the time being, don't replace-assign facts to themselves, as it is
very probably not handled right. Because of the current factstore
implementation this probably needs to be special-cased if we want to
support it.