Allow resolver variables to be forward-declared.
A forwad declaration simply interns the given variables to the symbol
table (giving them each a variable ID). Forward declarations are useful
when you have a plugin that resolves a hook-target for certain events
with a number of variables set, some of which might not be referenced
by all ruleset implementations. Without a forward-declaration the plugin
could not trigger the hook as the unreferenced variables would be unknown
and consequently unsettable.

The syntax of a forward declaration is "variables var1, var2, ... varN".
Forward declarations must be put after fact definitions and before any
targets definitions. You can use multiple forward declarations for
improved readability/logical grouping.