crash fix: pop src and dst from the stack in replacing assignments.
Always pop source and destination globals during the execution of
replacing assignments. Without this fix, in the case of no selector
fields, we might leave the processed and free'd source and destination
on the stack. Consequently if we run into a VM exception after a
replacing assignment we might crash or corrupt the heap on a double-free
during stack cleanup. For instance resolving this would crash:

test:
	$application_process *= fact('com.nokia.policy.application_process', \
	                             'process', 2567)
	fail()